+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+* `progress_handler` parameter to `wn.download()` ([#70])
+* `progress_handler` parameter to `wn.add()` ([#70])
+
 ### Changed
 
 * Renamed `lgcode` parameter to `lang` throughout ([#66])
@@ -60,3 +65,4 @@ abandoned, but this is an entirely new codebase.
 [#60]: https://github.com/goodmami/wn/issues/60
 [#61]: https://github.com/goodmami/wn/issues/61
 [#66]: https://github.com/goodmami/wn/issues/66
+[#70]: https://github.com/goodmami/wn/issues/70